groups the route actions together in package
