--- src/main/java/server/UnknownRoute.java
@@ -1,11 +1,0 @@
-package server;
-
-class UnknownRoute implements Action {
-    @Override
-    public HttpResponse process(HttpRequest request) {
-
-        return HttpResponseBuilder.anHttpResponseBuilder()
-                .withStatusCode(StatusCode.NOT_FOUND)
-                .build();
-    }
-}
