+++ src/main/java/server/actions/Redirect.java
@@ -1,0 +1,15 @@
+package server.actions;
+
+import server.*;
+
+public class Redirect implements Action {
+
+    @Override
+    public HttpResponse process(HttpRequest request) {
+
+        return HttpResponseBuilder.anHttpResponseBuilder()
+                .withStatusCode(StatusCode.FOUND)
+                .withLocation("http://localhost:5000/")
+                .build();
+    }
+}
